Fixed bad close(data->prev_fd) not being protected, resulting in close(0) for single commands and built_ins. Now exec behaves as Bash would
