+++ src/execution/redir_child.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   redir_child.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jteissie <jteissie@student.42.fr>          +#+  +:+       +#+        */
+/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/12 14:00:07 by jteissie          #+#    #+#             */
-/*   Updated: 2024/07/28 17:04:30 by jteissie         ###   ########.fr       */
+/*   Updated: 2024/07/28 20:55:30 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
