+++ src/signals.c
@@ -6,7 +6,7 @@
 /*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/11 12:57:23 by bthomas           #+#    #+#             */
-/*   Updated: 2024/07/25 17:51:25 by marvin           ###   ########.fr       */
+/*   Updated: 2024/07/28 19:57:18 by marvin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
